+++ SCrawler.PluginProvider/Interfaces/IPluginContentProvider.vb
@@ -26,6 +26,7 @@ Namespace Plugin
         Property SeparateVideoFolder As Boolean
         Property DataPath As String
         Property PostsNumberLimit As Integer?
+        Property PostsDateLimit As Date?
         Function ExchangeOptionsGet() As Object
         Sub ExchangeOptionsSet(ByVal Obj As Object)
         Sub XmlFieldsSet(ByVal Fields As List(Of KeyValuePair(Of String, String)))