+++ SCrawler/Editors/SiteEditorForm.Designer.vb
@@ -14,13 +14,13 @@
         Private components As System.ComponentModel.IContainer
         <System.Diagnostics.DebuggerStepThrough()>
         Private Sub InitializeComponent()
-            Dim ActionButton1 As PersonalUtilities.Forms.Controls.Base.ActionButton = New PersonalUtilities.Forms.Controls.Base.ActionButton()
+            Dim ActionButton7 As PersonalUtilities.Forms.Controls.Base.ActionButton = New PersonalUtilities.Forms.Controls.Base.ActionButton()
             Dim resources As System.ComponentModel.ComponentResourceManager = New System.ComponentModel.ComponentResourceManager(GetType(SiteEditorForm))
-            Dim ActionButton2 As PersonalUtilities.Forms.Controls.Base.ActionButton = New PersonalUtilities.Forms.Controls.Base.ActionButton()
-            Dim ActionButton3 As PersonalUtilities.Forms.Controls.Base.ActionButton = New PersonalUtilities.Forms.Controls.Base.ActionButton()
-            Dim ActionButton4 As PersonalUtilities.Forms.Controls.Base.ActionButton = New PersonalUtilities.Forms.Controls.Base.ActionButton()
-            Dim ActionButton5 As PersonalUtilities.Forms.Controls.Base.ActionButton = New PersonalUtilities.Forms.Controls.Base.ActionButton()
-            Dim ActionButton6 As PersonalUtilities.Forms.Controls.Base.ActionButton = New PersonalUtilities.Forms.Controls.Base.ActionButton()
+            Dim ActionButton8 As PersonalUtilities.Forms.Controls.Base.ActionButton = New PersonalUtilities.Forms.Controls.Base.ActionButton()
+            Dim ActionButton9 As PersonalUtilities.Forms.Controls.Base.ActionButton = New PersonalUtilities.Forms.Controls.Base.ActionButton()
+            Dim ActionButton10 As PersonalUtilities.Forms.Controls.Base.ActionButton = New PersonalUtilities.Forms.Controls.Base.ActionButton()
+            Dim ActionButton11 As PersonalUtilities.Forms.Controls.Base.ActionButton = New PersonalUtilities.Forms.Controls.Base.ActionButton()
+            Dim ActionButton12 As PersonalUtilities.Forms.Controls.Base.ActionButton = New PersonalUtilities.Forms.Controls.Base.ActionButton()
             Me.TP_MAIN = New System.Windows.Forms.TableLayoutPanel()
             Me.TXT_PATH = New PersonalUtilities.Forms.Controls.TextBoxExtended()
             Me.TXT_COOKIES = New PersonalUtilities.Forms.Controls.TextBoxExtended()
@@ -59,14 +59,14 @@
             '
             'TXT_PATH
             '
-            ActionButton1.BackgroundImage = CType(resources.GetObject("ActionButton1.BackgroundImage"), System.Drawing.Image)
-            ActionButton1.Index = 0
-            ActionButton1.Name = "BTT_OPEN"
-            ActionButton2.BackgroundImage = CType(resources.GetObject("ActionButton2.BackgroundImage"), System.Drawing.Image)
-            ActionButton2.Index = 1
-            ActionButton2.Name = "BTT_CLEAR"
-            Me.TXT_PATH.Buttons.Add(ActionButton1)
-            Me.TXT_PATH.Buttons.Add(ActionButton2)
+            ActionButton7.BackgroundImage = CType(resources.GetObject("ActionButton7.BackgroundImage"), System.Drawing.Image)
+            ActionButton7.Index = 0
+            ActionButton7.Name = "BTT_OPEN"
+            ActionButton8.BackgroundImage = CType(resources.GetObject("ActionButton8.BackgroundImage"), System.Drawing.Image)
+            ActionButton8.Index = 1
+            ActionButton8.Name = "BTT_CLEAR"
+            Me.TXT_PATH.Buttons.Add(ActionButton7)
+            Me.TXT_PATH.Buttons.Add(ActionButton8)
             Me.TXT_PATH.CaptionText = "Path"
             Me.TXT_PATH.CaptionToolTipEnabled = True
             Me.TXT_PATH.CaptionToolTipText = "Specific path to store Twitter files"
@@ -78,14 +78,14 @@
             '
             'TXT_COOKIES
             '
-            ActionButton3.BackgroundImage = CType(resources.GetObject("ActionButton3.BackgroundImage"), System.Drawing.Image)
-            ActionButton3.Index = 0
-            ActionButton3.Name = "BTT_EDIT"
-            ActionButton4.BackgroundImage = CType(resources.GetObject("ActionButton4.BackgroundImage"), System.Drawing.Image)
-            ActionButton4.Index = 1
-            ActionButton4.Name = "BTT_CLEAR"
-            Me.TXT_COOKIES.Buttons.Add(ActionButton3)
-            Me.TXT_COOKIES.Buttons.Add(ActionButton4)
+            ActionButton9.BackgroundImage = CType(resources.GetObject("ActionButton9.BackgroundImage"), System.Drawing.Image)
+            ActionButton9.Index = 0
+            ActionButton9.Name = "BTT_EDIT"
+            ActionButton10.BackgroundImage = CType(resources.GetObject("ActionButton10.BackgroundImage"), System.Drawing.Image)
+            ActionButton10.Index = 1
+            ActionButton10.Name = "BTT_CLEAR"
+            Me.TXT_COOKIES.Buttons.Add(ActionButton9)
+            Me.TXT_COOKIES.Buttons.Add(ActionButton10)
             Me.TXT_COOKIES.CaptionText = "Cookies"
             Me.TXT_COOKIES.ClearTextByButtonClear = False
             Me.TXT_COOKIES.Dock = System.Windows.Forms.DockStyle.Fill
@@ -114,14 +114,14 @@
             '
             'TXT_PATH_SAVED_POSTS
             '
-            ActionButton5.BackgroundImage = CType(resources.GetObject("ActionButton5.BackgroundImage"), System.Drawing.Image)
-            ActionButton5.Index = 0
-            ActionButton5.Name = "BTT_OPEN"
-            ActionButton6.BackgroundImage = CType(resources.GetObject("ActionButton6.BackgroundImage"), System.Drawing.Image)
-            ActionButton6.Index = 1
-            ActionButton6.Name = "BTT_CLEAR"
-            Me.TXT_PATH_SAVED_POSTS.Buttons.Add(ActionButton5)
-            Me.TXT_PATH_SAVED_POSTS.Buttons.Add(ActionButton6)
+            ActionButton11.BackgroundImage = CType(resources.GetObject("ActionButton11.BackgroundImage"), System.Drawing.Image)
+            ActionButton11.Index = 0
+            ActionButton11.Name = "BTT_OPEN"
+            ActionButton12.BackgroundImage = CType(resources.GetObject("ActionButton12.BackgroundImage"), System.Drawing.Image)
+            ActionButton12.Index = 1
+            ActionButton12.Name = "BTT_CLEAR"
+            Me.TXT_PATH_SAVED_POSTS.Buttons.Add(ActionButton11)
+            Me.TXT_PATH_SAVED_POSTS.Buttons.Add(ActionButton12)
             Me.TXT_PATH_SAVED_POSTS.CaptionText = "Saved posts path"
             Me.TXT_PATH_SAVED_POSTS.Dock = System.Windows.Forms.DockStyle.Fill
             Me.TXT_PATH_SAVED_POSTS.Location = New System.Drawing.Point(3, 31)